Исправлены недочеты: - Исправлена рамка индикации ответов - Исправлено множественное нажатие кнопок Да/Нет - Улучшена структура кода
