+++ MovieQuiz/Models/Model.swift
@@ -1,0 +1,18 @@
+import Foundation
+import UIKit
+
+struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
