+++ MovieQuiz/Presentation/Base.lproj/Main.storyboard
@@ -3,7 +3,7 @@
     <device id="retina6_1" orientation="portrait" appearance="dark"/>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21678"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21679"/>
         <capability name="Named colors" minToolsVersion="9.0"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
@@ -115,7 +115,9 @@
                     <connections>
                         <outlet property="counterLabel" destination="tve-bw-qfB" id="teO-OM-9SK"/>
                         <outlet property="imageView" destination="GyB-MK-pWj" id="nbv-Cn-sWR"/>
+                        <outlet property="noButton" destination="FvQ-vG-uSj" id="HRl-Ka-AGa"/>
                         <outlet property="textLabel" destination="bvA-xa-f0j" id="dda-7O-DfD"/>
+                        <outlet property="yesButton" destination="PgQ-MK-55e" id="8Ko-2B-7zS"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
